Abstracted reader into a util
