+++ docs/src/app/layout.tsx
@@ -2,14 +2,11 @@ import Head from 'next/head';
 import { Analytics } from '@vercel/analytics/react';
 
 import '../styles/global.css';
-import { createReader } from '@keystatic/core/reader';
-import keystaticConfig from '../../keystatic.config';
 import { HeaderNav } from '../components/navigation/header-nav';
-
-const reader = createReader('', keystaticConfig);
+import { getNavigationMap } from '../utils/reader';
 
 export const metadata = {
-  title: 'Meet Keystatic',
+  title: 'Keystatic',
   description:
     "Keystatic is a new tool from Thinkmill Labs that opens up your code-based content (written in Markdown, JSON or YAML) to contributors who aren't technical.",
   openGraph: {
@@ -36,25 +33,7 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const navigation = await reader.singletons.navigation.read();
-  const pages = await reader.collections.pages.all();
-
-  const pagesBySlug = Object.fromEntries(pages.map(page => [page.slug, page]));
-
-  const navigationMap = navigation?.navGroups.map(({ groupName, items }) => ({
-    groupName,
-    items: items.map(({ label, link }) => {
-      const { discriminant, value } = link;
-      const page = discriminant === 'page' && value ? pagesBySlug[value] : null;
-      const url = discriminant === 'url' ? value : `/docs/${page?.slug}`;
-
-      return {
-        label: label || page?.entry.title || '',
-        href: url || '',
-        title: page?.entry.title,
-      };
-    }),
-  }));
+  const navigationMap = await getNavigationMap();
 
   return (
     <html lang="en">
@@ -85,7 +64,6 @@ export default async function RootLayout({
       <body>
         <div className="min-h-screen">
           <HeaderNav navigationMap={navigationMap} />
-
           {children}
         </div>
       </body>
